delete comment from post comment array

+++ controllers/commentController.js
@@ -5,9 +5,9 @@ const Comments = require('../models/commentModel');
 const commentController = {
     createComment: async(req,res) => {
        try{
-            const {postId, content, tag, reply} = req.body;
+            const {postId, content, tag, reply,postUserId} = req.body;
             const newComment = new Comments({
-                user: req.user._id, content, tag, reply
+                user: req.user._id, content, tag, reply, postUserId, postId
             })
 
             await Posts.findOneAndUpdate({_id: postId}, {
@@ -55,8 +55,12 @@ const commentController = {
     },
     deleteComment: async(req,res) => {
         try{
-            const comment = await Comments.findOneAndDelete({_id: req.params.id});
+            const comment = await Comments.findOneAndDelete({_id: req.params.id, $or: [{user: req.user._id},{postUserId: req.user._id}]});
+            await Posts.findOneAndUpdate({_id: comment.Id}, {
+                $pull: {comments: req.params.id}
+            })
             if(comment.err) return res.status(400).json({msg: post.err});
+        
             res.json({msg: "You Deleted your comment"});
         } catch(err) {
             return res.status(500).json({msg: err.message});
